requirements.txt
Django==2.2.12
django-bootstrap4==1.1.1
django-tinymce==3.0.2